api/scnplusrmm/requirements.txt
asgiref==3.8.1
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
cryptography==44.0.2
Django==4.2.22
django-cors-headers==4.7.0
django-allauth[socialaccount]==65.2.0
django-filter==25.1
django-rest-knox==4.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
hiredis==2.3.2
kombu==5.4.2
meshctrl==0.1.15
msgpack==1.1.0
nats-py==2.10.0
packaging==24.2
psutil==7.0.0
psycopg[binary]==3.2.9
pycparser==2.22
pycryptodome==3.22.0
pyotp==2.9.0
pyparsing==3.2.3
python-ipware==2.0.2
qrcode==8.2
redis==5.0.8
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tldextract==5.1.3
twilio==8.13.0
urllib3==2.3.0
uvicorn[standard]==0.34.3
uWSGI==2.0.30
validators==0.24.0
vine==5.1.0
websockets==13.1
zipp==3.21.0
pandas==2.2.3
kaleido==0.2.1
jinja2==3.1.6
markdown==3.7
plotly==5.24.1
weasyprint==64.1
ocxsect==0.1.5
